Operations refactoring by kafka message type
